refactor(message-domain) : Message Entity 추가
